*** empty log message ***

+++ testing/test.sh
@@ -5,6 +5,11 @@ set refin=../testing/reference2d.inp
 set refout=../testing/reference2d.out
 set tmp=/tmp/temp.out
 
+if ($#argv == 0 ) then
+   echo "./test.sh [makeref,twod,threed]"
+   exit
+endif
+
 if ($1 == makeref) then
 
    ./gridsetup.py 1 1 1 128 128 1 
@@ -12,19 +17,25 @@ if ($1 == makeref) then
    ./dns < $refin > $refout
   cat $refout
 
-else
+endif
 
-./gridsetup.py 1 1 1 128 128 6  0 0 0  2 2 2
+if ($1 == threed) then
+
+  ./gridsetup.py 1 1 1 128 128 6  0 0 0  2 2 2
+  make >& /dev/null ;  rm -f $tmp ; ./dns < $refin > $tmp ; diff $tmp $refout 
+
+  ./gridsetup.py 1 1 1 128 128 6  3 2 1  5 4 3
+  make >& /dev/null ;  rm -f $tmp ; ./dns < $refin > $tmp ; diff $tmp $refout 
+
+  ./gridsetup.py 1 1 1 128 128 6  2 2 2  0 0 0 
+  make >& /dev/null ;  rm -f $tmp ; ./dns < $refin > $tmp ; diff $tmp $refout 
+
+endif
+
+if ($1 == twod) then
+
+./gridsetup.py 1 1 1 128 128 1  0 0 0  2 2 0
 make >& /dev/null ;  rm -f $tmp ; ./dns < $refin > $tmp ; diff $tmp $refout
-
-./gridsetup.py 1 1 1 128 128 6  3 2 1  5 4 3
-make >& /dev/null ;  rm -f $tmp ; ./dns < $refin > $tmp ; diff $tmp $refout
-
-./gridsetup.py 1 1 1 128 128 6  2 2 2  0 0 0 
-make >& /dev/null ;  rm -f $tmp ; ./dns < $refin > $tmp ; diff $tmp $refout
-
-
-
 
 ./gridsetup.py 1 1 1 128 128 1  2 2 2  0 0 0 
 make >& /dev/null ;  rm -f $tmp ; ./dns < $refin > $tmp ; diff $tmp $refout
@@ -39,8 +50,6 @@ make >& /dev/null ;  rm -f $tmp ; ./dns < $refin > $tmp ; diff $tmp $refout
 ./gridsetup.py 1 1 1 128 128 1  4 2 3  0 0 0 
 make >& /dev/null ;  rm -f $tmp ; ./dns < $refin > $tmp ; diff $tmp $refout
 
-./gridsetup.py 1 1 1 128 128 1  0 0 0  2 2 0
-make >& /dev/null ;  rm -f $tmp ; ./dns < $refin > $tmp ; diff $tmp $refout
 
 ./gridsetup.py 1 1 1 128 128 1  0 0 0  2 3 1
 make >& /dev/null ;  rm -f $tmp ; ./dns < $refin > $tmp ; diff $tmp $refout
